test run

+++ src/abiEncoderV2/byte_arrays.sol
@@ -1,13 +1,19 @@
 pragma abicoder v2;
 
 contract C {
-    function f(uint a, bytes memory b, uint c)
-            public pure returns (uint, uint, bytes1, uint) {
+    function f(uint256 a, bytes memory b, uint256 c)
+        public
+        pure
+        returns (uint256, uint256, bytes1, uint256)
+    {
         return (a, b.length, b[3], c);
     }
 
-    function f_external(uint a, bytes calldata b, uint c)
-            external pure returns (uint, uint, bytes1, uint) {
+    function f_external(uint256 a, bytes calldata b, uint256 c)
+        external
+        pure
+        returns (uint256, uint256, bytes1, uint256)
+    {
         return (a, b.length, b[3], c);
     }
 }
