+++ src/abiEncoderV2/memory_params_in_external_function.sol
@@ -1,15 +1,19 @@
 pragma abicoder v2;
 
 contract C {
-    function f(bytes memory a, bytes calldata b, uint[] memory c)
+    function f(bytes memory a, bytes calldata b, uint256[] memory c)
         external
         pure
-        returns (uint, bytes1, uint, bytes1, uint, uint)
+        returns (uint256, bytes1, uint256, bytes1, uint256, uint256)
     {
         return (a.length, a[1], b.length, b[2], c.length, c[3]);
     }
-    function g() public returns (uint, bytes1, uint, bytes1, uint, uint) {
-        uint[] memory x = new uint[](4);
+
+    function g()
+        public
+        returns (uint256, bytes1, uint256, bytes1, uint256, uint256)
+    {
+        uint256[] memory x = new uint[](4);
         x[3] = 7;
         return this.f("abc", "def", x);
     }
