+++ src/array/calldata_bytes_array_bounds.sol
@@ -1,6 +1,7 @@
-pragma abicoder               v2;
+pragma abicoder v2;
+
 contract C {
-    function f(bytes[] calldata a, uint256 i) external returns (uint) {
+    function f(bytes[] calldata a, uint256 i) external returns (uint256) {
         return uint8(a[0][i]);
     }
 }
