+++ src/array/dynamic_multi_array_cleanup.sol
@@ -1,17 +1,21 @@
 contract c {
-    struct s { uint[][] d; }
+    struct s {
+        uint256[][] d;
+    }
+
     s[] data;
-    function fill() public returns (uint) {
-        while (data.length < 3)
-            data.push();
-        while (data[2].d.length < 4)
-            data[2].d.push();
-        while (data[2].d[3].length < 5)
-            data[2].d[3].push();
+
+    function fill() public returns (uint256) {
+        while (data.length < 3) data.push();
+        while (data[2].d.length < 4) data[2].d.push();
+        while (data[2].d[3].length < 5) data[2].d[3].push();
         data[2].d[3][4] = 8;
         return data[2].d[3][4];
     }
-    function clear() public { delete data; }
+
+    function clear() public {
+        delete data;
+    }
 }
 // ----
 // storageEmpty -> 1
