+++ src/array/length/fixed_size_zero_length.sol
@@ -1,10 +1,10 @@
 contract C {
-  int[0] a;
-  uint[0] b;
-  bytes1[0] c;
-  bytes32[0] d;
-  bytes[0] e;
-  string[0] f;
+    int256[0] a;
+    uint256[0] b;
+    bytes1[0] c;
+    bytes32[0] d;
+    bytes[0] e;
+    string[0] f;
 }
 // ----
 // TypeError 1406: (19-20): Array with zero length specified.
