+++ src/array/no_array_pop.sol
@@ -1,7 +1,8 @@
 contract C {
-    uint data;
+    uint256 data;
+
     function test() public {
-      data.pop();
+        data.pop();
     }
 }
 // ----
