+++ src/array/pop/array_pop_storage_empty.sol
@@ -1,5 +1,6 @@
 contract c {
-    uint[] data;
+    uint256[] data;
+
     function test() public {
         data.push(7);
         data.pop();
