+++ src/array/uninitialized_storage_var.sol
@@ -1,10 +1,10 @@
 contract C {
-	function f() public {
-		uint[] storage x;
-		uint[10] storage y;
-		x;
-		y;
-	}
+    function f() public {
+        uint256[] storage x;
+        uint256[10] storage y;
+        x;
+        y;
+    }
 }
 // ----
 // TypeError 3464: (80-81): This variable is of storage pointer type and can be accessed without prior assignment, which would lead to undefined behaviour.
