+++ src/calldata/calldata_bound_dynamic_array_or_slice.sol
@@ -1,7 +1,11 @@
 pragma abicoder v2;
 
 library L {
-    function reverse(uint[] calldata _a) internal pure returns (uint, uint) {
+    function reverse(uint256[] calldata _a)
+        internal
+        pure
+        returns (uint256, uint256)
+    {
         return (_a[1], _a[0]);
     }
 }
@@ -9,15 +13,22 @@ library L {
 contract C {
     using L for *;
 
-    function testArray(uint, uint[] calldata _a, uint) external pure returns (uint, uint) {
+    function testArray(uint256, uint256[] calldata _a, uint256)
+        external
+        pure
+        returns (uint256, uint256)
+    {
         return _a.reverse();
     }
 
-    function testSlice(uint, uint[] calldata _a, uint) external pure returns (uint, uint) {
+    function testSlice(uint256, uint256[] calldata _a, uint256)
+        external
+        pure
+        returns (uint256, uint256)
+    {
         return _a[:].reverse();
     }
 }
-
 // ----
 // testArray(uint256,uint256[],uint256): 7, 0x60, 4, 2, 66, 77 -> 77, 66
 // testSlice(uint256,uint256[],uint256): 7, 0x60, 4, 2, 66, 77 -> 77, 66
