+++ src/calldata/calldata_bound_static_array.sol
@@ -1,18 +1,25 @@
 pragma abicoder v2;
 
 library L {
-    function reverse(uint[2] calldata _a) internal pure returns (uint, uint) {
+    function reverse(uint256[2] calldata _a)
+        internal
+        pure
+        returns (uint256, uint256)
+    {
         return (_a[1], _a[0]);
     }
 }
 
 contract C {
-    using L for uint[2];
+    using L for uint256[2];
 
-    function test(uint, uint[2] calldata _a, uint) external pure returns (uint, uint) {
+    function test(uint256, uint256[2] calldata _a, uint256)
+        external
+        pure
+        returns (uint256, uint256)
+    {
         return _a.reverse();
     }
 }
-
 // ----
 // test(uint256,uint256[2],uint256): 7, 66, 77, 4 -> 77, 66
