+++ src/functionCall/calling_other_functions.sol
@@ -1,14 +1,15 @@
 contract collatz {
-    function run(uint x) public returns(uint y) {
+    function run(uint256 x) public returns (uint256 y) {
         while ((y = x) > 1) {
-            if (x % 2 == 0) x = evenStep(x);
-            else x = oddStep(x);
+            if (x % 2 == 0) x = evenStep(x); else x = oddStep(x);
         }
     }
-    function evenStep(uint x) public returns(uint y) {
+
+    function evenStep(uint256 x) public returns (uint256 y) {
         return x / 2;
     }
-    function oddStep(uint x) public returns(uint y) {
+
+    function oddStep(uint256 x) public returns (uint256 y) {
         return 3 * x + 1;
     }
 }
