+++ src/getter/complex_struct.sol
@@ -1,8 +1,8 @@
 contract C {
     struct Y {
-        uint a;
-        uint b;
+        uint256 a;
+        uint256 b;
     }
+
     mapping(uint256 => Y) public m;
-}
-// ----
+} // ----
