+++ src/immutable/ctor_initialization_indirect_reading.sol
@@ -1,10 +1,13 @@
 contract C {
-    uint immutable x;
+    uint256 immutable x;
+
     constructor() {
         x = f();
     }
 
-    function f() public pure returns (uint) { return 3 + x; }
+    function f() public pure returns (uint256) {
+        return 3 + x;
+    }
 }
 // ----
 // TypeError 7733: (136-137): Immutable variables cannot be read before they are initialized.
