+++ src/immutable/function_initialization_reading.sol
@@ -1,7 +1,9 @@
 contract C {
-    uint immutable x = f();
+    uint256 immutable x = f();
 
-    function f() public pure returns (uint) { return 3 + x; }
+    function f() public pure returns (uint256) {
+        return 3 + x;
+    }
 }
 // ----
 // TypeError 7733: (99-100): Immutable variables cannot be read before they are initialized.
