+++ src/immutable/increment_decrement.sol
@@ -1,10 +1,11 @@
 contract C {
-  uint immutable x;
-  uint immutable y;
-  constructor() {
-    ++x;
-    --y;
-  }
+    uint256 immutable x;
+    uint256 immutable y;
+
+    constructor() {
+        ++x;
+        --y;
+    }
 }
 // ----
 // TypeError 3969: (77-78): Immutable variables must be initialized using an assignment.
