+++ src/immutable/indirect_reading_during_statevar_init.sol
@@ -1,8 +1,10 @@
 contract C {
-    uint immutable x = 0;
-    uint y = f();
+    uint256 immutable x = 0;
+    uint256 y = f();
 
-    function f() internal returns(uint) { return x; }
+    function f() internal returns (uint256) {
+        return x;
+    }
 }
 // ----
 // TypeError 7733: (107-108): Immutable variables cannot be read before they are initialized.
