+++ src/immutable/inheritance_virtual_functions.sol
@@ -1,17 +1,17 @@
 contract B {
-    uint immutable x;
+    uint256 immutable x;
 
     constructor() {
         x = xInit();
     }
 
-    function xInit() internal virtual returns(uint) {
+    function xInit() internal virtual returns (uint256) {
         return 3;
     }
 }
 
 contract C is B {
-    function xInit() internal override returns(uint) {
+    function xInit() internal override returns (uint256) {
         return x;
     }
 }
