+++ src/immutable/multi_creation.sol
@@ -1,29 +1,41 @@
 contract A {
-	uint immutable a;
-	constructor() {
-		a = 7;
-	}
-	function f() public view returns (uint) { return a; }
+    uint256 immutable a;
+
+    constructor() {
+        a = 7;
+    }
+
+    function f() public view returns (uint256) {
+        return a;
+    }
 }
+
 contract B {
-	uint immutable a;
-	constructor() {
-		a = 5;
-	}
-	function f() public view returns (uint) { return a; }
+    uint256 immutable a;
+
+    constructor() {
+        a = 5;
+    }
+
+    function f() public view returns (uint256) {
+        return a;
+    }
 }
+
 contract C {
-	uint immutable a;
-	uint public x;
-	uint public y;
-	constructor() {
-		a = 3;
-		x = (new A()).f();
-		y = (new B()).f();
-	}
-	function f() public returns (uint256, uint, uint) {
-		return (a, (new A()).f(), (new B()).f());
-	}
+    uint256 immutable a;
+    uint256 public x;
+    uint256 public y;
+
+    constructor() {
+        a = 3;
+        x = (new A()).f();
+        y = (new B()).f();
+    }
+
+    function f() public returns (uint256, uint256, uint256) {
+        return (a, (new A()).f(), (new B()).f());
+    }
 }
 // ----
 // f() -> 3, 7, 5
