+++ src/immutable/unrelated_reading.sol
@@ -1,8 +1,7 @@
 contract C {
-    uint immutable x = 1;
+    uint256 immutable x = 1;
 
-    function readX() internal pure returns(uint) {
+    function readX() internal pure returns (uint256) {
         return x + 3;
     }
-}
-// ----
+} // ----
