+++ src/nameAndTypeResolution/142_inheritence_suggestions.sol
@@ -1,7 +1,10 @@
-contract a { function func() public {} }
+contract a {
+    function func() public {}
+}
+
 contract c is a {
     function g() public {
-        uint var1 = fun();
+        uint256 var1 = fun();
     }
 }
 // ----
