+++ src/nameAndTypeResolution/220_struct_constructor.sol
@@ -1,5 +1,9 @@
 contract C {
-    struct S { uint a; bool x; }
+    struct S {
+        uint256 a;
+        bool x;
+    }
+
     function f() public {
         S memory s = S(1, true);
     }
