+++ src/nameAndTypeResolution/474_storage_location_non_array_or_struct_disallowed_is_not_fatal.sol
@@ -1,5 +1,5 @@
 contract C {
-    function f(uint storage a) public {
+    function f(uint256 storage a) public {
         a = f;
     }
 }
