+++ src/natspec/docstring_empty_tag.sol
@@ -1,6 +1,6 @@
 abstract contract C {
     /// @param
-    function vote(uint id) public {}
+    function vote(uint256 id) public {}
 }
 // ----
 // DocstringParsingError 3335: (26-36): No param name given
