+++ src/revertStrings/invalid_abi_decoding_memory_v1.sol
@@ -1,16 +1,21 @@
 pragma abicoder v1;
+
 contract C {
-	function dyn(uint ptr, uint start, uint x) public returns (bytes memory a) {
-		assembly {
-			mstore(0, start)
-			mstore(start, add(start, 1))
-			return(ptr, x)
-		}
-	}
-	function f(uint ptr, uint start, uint x) public returns (bool) {
-		this.dyn(ptr, start, x);
-		return true;
-	}
+    function dyn(uint256 ptr, uint256 start, uint256 x)
+        public
+        returns (bytes memory a)
+    {
+        assembly {
+    mstore(0, start)
+    mstore(start, add(start, 1))
+    return(ptr, x)
+    }
+    }
+
+    function f(uint256 ptr, uint256 start, uint256 x) public returns (bool) {
+        this.dyn(ptr, start, x);
+        return true;
+    }
 }
 // ====
 // EVMVersion: >=byzantium
