+++ src/structs/array_calldata.sol
@@ -1,7 +1,10 @@
-pragma abicoder               v2;
+pragma abicoder v2;
+
 contract Test {
-    struct S { int a; }
-    function f(S[] calldata) external { }
-    function f(S[][] calldata) external { }
-}
-// ----
+    struct S {
+        int256 a;
+    }
+
+    function f(S[] calldata) external {}
+    function f(S[][] calldata) external {}
+} // ----
