+++ src/types/address/address_payable_internal_overload_nonpayable.sol
@@ -1,10 +1,10 @@
 contract C {
     function f(address payable) internal pure {}
-    function f(address) internal pure returns (uint) {}
+    function f(address) internal pure returns (uint256) {}
+
     function g() internal pure {
         address a = address(0);
-        uint b = f(a); // TODO: should this be valid?
+        uint256 b = f(a); // TODO: should this be valid?
         b;
     }
-}
-// ----
+} // ----
